refacto(authbundle): split code.
