+++ bundles/authenticator/token.go
@@ -2,7 +2,6 @@ package authenticator
 
 import (
 	"errors"
-	"io/ioutil"
 	"strconv"
 	"time"
 
@@ -10,13 +9,6 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-// Rsa is the struct to get the rsa keys used to generate and verify tokens from the environment variables
-type Rsa struct {
-	PublicKeyPath  string
-	PrivateKeyPath string
-	PublicKey      interface{}
-	PrivateKey     interface{}
-}
 
 // Claim is the struct for the jwt claim
 type Claim struct {
@@ -24,50 +16,12 @@ type Claim struct {
 	jwt.StandardClaims
 }
 
-var rsa Rsa
-
-func initRsaKeys() error {
-	if rsa.PrivateKey != nil && rsa.PublicKey != nil {
-		return nil
-	}
-
-	rsa.PublicKeyPath = goDotEnvVariable("RSA_PUBLIC_PATH")
-	rsa.PrivateKeyPath = goDotEnvVariable("RSA_PRIVATE_PATH")
-
-	// Get the public key
-	publicKeyData, err := ioutil.ReadFile(rsa.PublicKeyPath)
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-	publicKey, err := jwt.ParseRSAPublicKeyFromPEM(publicKeyData)
-	if err != nil {
-		log.Error("error public key: ", err)
-		return err
-	}
-	rsa.PublicKey = publicKey
-
-	// Get the private key
-	privateKeyData, err := ioutil.ReadFile(rsa.PrivateKeyPath)
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-	privateKey, err := jwt.ParseRSAPrivateKeyFromPEMWithPassword(privateKeyData, goDotEnvVariable("RSA_PASSWORD"))
-	if err != nil {
-		log.Error("error public key: ", err)
-		return err
-	}
-	rsa.PrivateKey = privateKey
-
-	return nil
-}
 
 // GenerateToken creates a JWT with email and expiration time in the payload
 func GenerateToken(email string) (string, error) {
 	err := initRsaKeys()
 	if err != nil {
-		return "", errors.New("Couldn't init rsa keys")
+		return "", errors.New("couldn't init rsa keys")
 	}
 
 	validTime, _ := strconv.ParseInt(goDotEnvVariable("TOKEN_VALID_DURATION"), 10, 64)
@@ -97,7 +51,7 @@ func GenerateToken(email string) (string, error) {
 func DecodeToken(tokenString string) (*jwt.Token, *Claim, error) {
 	err := initRsaKeys()
 	if err != nil {
-		return nil, &Claim{}, errors.New("Couldn't init rsa keys")
+		return nil, &Claim{}, errors.New("couldn't init rsa keys")
 	}
 
 	claims := &Claim{}
